+++ src/relstorage/zodbconvert.py
@@ -106,7 +106,8 @@ def main(argv=None):
 
     logging.basicConfig(
         level=logging.INFO,
-        format="%(asctime)s [%(name)s] %(levelname)s %(message)s")
+        format="%(asctime)s [%(name)s] %(levelname)s %(message)s"
+    )
 
     source, destination = open_storages(options)